appscripts: add user friendly alarm menu
Adds a new script to manage alarms with a user friendly menu.
It uses user cron jobs and mpv for music playback.

The entries generated in user's crontab are separated by two
"guard" comments, other entries outside of those will not be touched.

The sound for the alarm can be placed under .config/sxmo/alarm.mp3,
or can be moved to a custom place and specified in the environment
variable ALARM_PATH. If none of them exists, it will use the default
ring tone that phone call is using.

The volume and output device can be set through ALARM_VOLUME
(defaults to 30), and ALARM_AUDIOOUT (defaults to Speaker).