Don't start a dbus session if one is already running
Systemd starts a dbus session bus automatically, and runs user services
on it. When we create our own session bus, we can't talk to those
services.

An alternative method of implementing this would be to inform systemd of
the of the bus we start, but I don't see why we would want to run extra
bus sessions.