Disable most of input handler when a locker is running
A problem was introduced in 4562daea673c6d936e2fdf9adc97984f84bd0d86
because it kind-of bypasses the state mechanism. Whenever a locker was
running the state would remain `unlock` and therefore the inputhandler
would be fully enabled and menus could still be opened with gestured
(although they wouldn't show because of the locker).

This patch solves that, it's not an ideal implementation but is
a bit of a minimal temporary patch until we revise/simplify the
entire state locking mechanism.

It also moves the burden of removing icons for peanutbutter to
sxmo_hook_statusbar instead of sxmo_hook_lockstatusbar

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>